warning cleanup: second pass in "data" directory
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@23706 18c085ea-50e0-402c-830e-de6fd14e8384

+++ data/bounce.C
@@ -413,7 +413,7 @@ public:
                     double x[nvals];
                     double y[nvals];
                     double z[nvals];
-                    for(size_t i = 0; i < nvals; ++i)
+                    for(int i = 0; i < nvals; ++i)
                     {
                         double t = double(i) / double(nvals-1);
                         x[i] = (1.-t)*thisExtents[0] + t*thisExtents[1];
